WikiaLogger cleanup: moved hook to own class
The WikiaLogger class included a hook function for loading some JS on
the page to handle 'onerror'. This commit moves that code to it's own
class and adds the hook in DefaultSettings.php along with the other
'WikiaSkinTopScripts' hooks.
